Extract NetworkOptions internal structure to core package
